test sensors

+++ components/iq2020/iq2020.cpp
@@ -466,10 +466,10 @@ int IQ2020Component::processIQ2020Command() {
 			//int UnknownCounter4 = readCounter(data, 86);
 
 			// Mystery values
-			if (this->testval1_sensor_) this->testval1_sensor_->publish_state(processingBuffer[98]);
-			if (this->testval2_sensor_) this->testval2_sensor_->publish_state(processingBuffer[100]);
-			if (this->testval3_sensor_) this->testval3_sensor_->publish_state(processingBuffer[116]);
-			if (this->testval4_sensor_) this->testval4_sensor_->publish_state(processingBuffer[128]);
+			if (this->testval1_sensor_) this->testval1_sensor_->publish_state((float)processingBuffer[98]);
+			if (this->testval2_sensor_) this->testval2_sensor_->publish_state((float)processingBuffer[100]);
+			if (this->testval3_sensor_) this->testval3_sensor_->publish_state((float)processingBuffer[116]);
+			if (this->testval4_sensor_) this->testval4_sensor_->publish_state((float)processingBuffer[128]);
 #endif
 
 			if (pending_temp != -1) {
